disabled input, moved actionbar changes to .tsx
