+++ src/frontend/styles/Game.module.css
@@ -74,6 +74,7 @@ top: 580.41px; */
   height: 44px;
   color: #002a52;
   font-size: 26px;
+  text-align: center;
 }
 
 .actionBarRight {
